+++ Subsurface/Source/Events/Missions/Mission.cs
@@ -230,7 +230,7 @@ namespace Barotrauma
 
         public virtual void Update(float deltaTime) { }
 
-        public virtual bool AssignTeamIDs(List<Networking.Client> clients, out int hostTeam)
+        public virtual bool AssignTeamIDs(List<Networking.Client> clients, out byte hostTeam)
         {
             clients.ForEach(c => c.TeamID = 1);
             hostTeam = 1;
